feature/views The views were completed
