+++ src/components/atoms/Logo.jsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import logo from '../../assets/logo.png';
+import { Link } from 'react-router-dom';
 
 const Logo = () => {
   return (
-    <img src={ logo } alt="Logo" />
+    <Link to="/">
+      <img src={ logo } alt="Logo" />
+    </Link>
   );
 };
