+++ src/components/molecules/BabysittersMap.jsx
@@ -7,8 +7,8 @@ const BabysittersMap = compose(
   withProps({
     googleMapURL: 'https://maps.googleapis.com/maps/api/js?key=AIzaSyAeTBVGhPqkbeNcVh-sEXG5K_l4CBhzZCQ&v=3.exp&libraries=geometry,drawing,places',
     loadingElement: <div style={ { height: `100%` } } />,
-    containerElement: <div style={ { height: `400px` } } />,
-    mapElement: <div style={ { height: `100%`, width: `100%` } } />,
+    containerElement: <div style={ { height: `100%` } } />,
+    mapElement: <div style={ { height: `100%`, width: `100%`, borderTopRightRadius: `20px`, borderBottomRightRadius: `20px` } } />,
   }),
   withScriptjs,
   withGoogleMap
