+++ src/styles/components/molecules/BabysitterMiniCard.scss
@@ -2,7 +2,6 @@
   display: inline-flex;
   align-items: center;
   padding: $space-small;
-  border: $color-Black solid 1px;
   text-decoration: none;
   color: $color-Black;
   img {
@@ -15,6 +14,7 @@
   }
   .babysitterMiniCard-info {
     h4 {
+      width: 220px;
       margin-bottom: $space-nano;
       text-align: left;
     }
