+++ src/styles/components/molecules/ImageInput.scss
@@ -1,5 +1,8 @@
 .imageInput {
   position: relative;
+  width: 200px;
+  height: 200px;
+  box-sizing: border-box;
   &-inputfile {
     width: 0.1px;
     height: 0.1px;
@@ -9,22 +12,26 @@
     z-index: -1;
   }
   &-inputfile + label {
+    position: absolute;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
     width: 100%;
+    height: 100%;
     display: inline-block;
-    padding: 96px 35px;
-    border: solid 1px $color-Black;
+    padding: 72px 35px;
     border-radius: 20px;
-    background-color: $color-Blue-Background;
     color: $color-Black;
     cursor: pointer;
     box-sizing: border-box;
-    @include h2;
+    font-size: 28px;
+    font-family: $font-Family-Overlock;
+    @include container($color-Brown-Main);
   }
   &-preview {
     display: none;
     width: 100%;
-    height: 232px;
-    border: 1px $color-Black solid;
+    height: 100%;
     border-radius: 20px;
     top: 0;
     left: 0;
