+++ src/styles/containers/clients/BabysitterDetail.scss
@@ -1,0 +1,69 @@
+.babysitterDetail {
+  &-container {
+    padding: $space-medium 0;
+  }
+  &-details {
+    display: flex;
+    align-items: center;
+    padding: $space-small;
+    @include container($color-Background-Container);
+    .image-container {
+      margin-right: #{$space-small * 3};
+      @include picture(200px);
+    }
+    .details-info {
+      text-align: left;
+      & > * {
+        margin-bottom: $space-medium;
+      }
+      & > *:last-child {
+        margin-bottom: 0;
+      }
+    }
+  }
+  section {
+    margin-bottom: $space-medium;
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+  &-about {
+    display: grid;
+    grid-template-columns: 70% 30%;
+    padding: $space-small #{$space-small * 4};
+    background-color: $color-Background-Section;
+    .about-container {
+      margin-right: $space-medium;
+    }
+    .price {
+      height: fit-content;
+      padding: $space-small $space-medium;
+      @include container($color-Background-Container);
+      background-color: $color-Brown-Main;
+      h3 {
+        margin-bottom: $space-small;
+      }
+    }
+    .textArea {
+      margin-bottom: $space-small;
+      textarea {
+        background-color: $color-Brown-Main;
+      }
+    }
+  }
+  &-availability {
+    .availability-container {
+      display: flex;
+      flex-flow: column;
+      align-items: center;
+      width: fit-content;
+      padding: $space-medium;
+      margin: 0 auto;
+      @include container($color-Background-Container);
+      h4 {
+        margin-bottom: $space-small;
+      }
+    }
+  }
+
+}
