+++ src/styles/containers/clients/PersonalInfoForm.scss
@@ -1,0 +1,51 @@
+.personalInfoForm {
+  padding: $space-medium;
+  &-container {
+    display: block;
+    padding: $space-medium;
+    @include container($color-Background-Container);
+    h2 {
+      margin-bottom: $space-medium;
+    }
+    form {
+      display: grid;
+      grid-template-columns: 60% 40%;
+      .left {
+        display: flex;
+        flex-flow: column;
+        align-items: flex-start;
+        margin-right: $space-medium;
+        div {
+          margin-bottom: $space-medium;
+        }
+      }
+      .right {
+        display: flex;
+        flex-flow: column;
+        align-items: center;
+        div {
+          margin-bottom: $space-medium;
+        }
+      }
+      & > div.input {
+        width: 100%;
+        grid-column-start: 1;
+        grid-column-end: -1;
+        margin-bottom: $space-medium;
+        input {
+          width: 80%;
+        }
+      }
+      button {
+        grid-column-start: 1;
+        grid-column-end: -1;
+        justify-self: center;
+      }
+      strong {
+        font-size: 18px;
+        font-family: $font-Family-Roboto;
+        color: $color-Negative;
+      }
+    }
+  }
+}
