+++ src/styles/main.scss
@@ -54,10 +54,10 @@
  **/
 
 @import "./containers/Login.scss";
-@import "./containers/clients/Register.scss";
+@import "./containers/clients/PersonalInfoForm.scss";
 @import "./containers/Email.scss";
 @import "./containers/clients/Babysitters.scss";
-@import "./containers/clients/Babysitter.scss";
+@import "./containers/clients/BabysitterDetail.scss";
 @import "./containers/clients/ServiceForm.scss";
 @import "./containers/clients/ServiceResume.scss";
 @import "./containers/clients/Schedule.scss";
